Add files via upload
The system can see the request of the adopters and has options whether to reject or approve, and if the adopter is being accepted or approved, it will proceed to the schedule of the interview. It is notified through the send of an email, and if the adopter is rejected, the pet remains available for adoption.